Added confirmation dialog for deletion operations. Fixed import of XML data causing crashes or not updating UI properly.

+++ att-ui/src/main/java/org/zlogic/att/ui/CustomFieldEditorController.java
@@ -6,7 +6,11 @@
 package org.zlogic.att.ui;
 
 import java.net.URL;
+import java.text.MessageFormat;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.ResourceBundle;
+import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
@@ -14,6 +18,8 @@ import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.TextFieldTableCell;
+import javafx.scene.image.Image;
+import javafx.stage.Stage;
 import javafx.util.Callback;
 import javafx.util.converter.DefaultStringConverter;
 import org.zlogic.att.ui.adapters.CustomFieldAdapter;
@@ -27,6 +33,10 @@ import org.zlogic.att.ui.adapters.DataManager;
  */
 public class CustomFieldEditorController implements Initializable {
 
+	/**
+	 * Localization messages
+	 */
+	private static final ResourceBundle messages = ResourceBundle.getBundle("org/zlogic/att/ui/messages");
 	/**
 	 * DataManager reference
 	 */
@@ -46,6 +56,10 @@ public class CustomFieldEditorController implements Initializable {
 	 */
 	@FXML
 	private TableView<CustomFieldAdapter> customFields;
+	/**
+	 * Confirmation prompt dialog controller
+	 */
+	private ConfirmationDialogController confirmationDialogController;
 
 	/**
 	 * Initializes the controller
@@ -68,6 +82,20 @@ public class CustomFieldEditorController implements Initializable {
 				return cell;
 			}
 		});
+
+		//Confirmation dialog
+		confirmationDialogController = ConfirmationDialogController.createInstance();
+	}
+
+	/**
+	 * Sets the window icons
+	 *
+	 * @param icons the icons to be set
+	 */
+	public void setWindowIcons(ObservableList<Image> icons) {
+		if (deleteButton.getScene().getWindow() instanceof Stage)
+			((Stage) deleteButton.getScene().getWindow()).getIcons().setAll(icons);
+		confirmationDialogController.setWindowIcons(icons);
 	}
 
 	/**
@@ -106,7 +134,16 @@ public class CustomFieldEditorController implements Initializable {
 	 */
 	@FXML
 	private void deleteCustomField() {
-		for (CustomFieldAdapter customField : customFields.getSelectionModel().getSelectedItems())
-			dataManager.deleteCustomField(customField);
+		StringBuilder tasksToDelete = new StringBuilder();
+		List<CustomFieldAdapter> customfieldsToDeleteList = new LinkedList(customFields.getSelectionModel().getSelectedItems());
+		for (CustomFieldAdapter customfieldToDelete : customfieldsToDeleteList)
+			tasksToDelete.append(tasksToDelete.length() > 0 ? "\n" : "").append(customfieldToDelete.nameProperty().get()); //NOI18N
+		ConfirmationDialogController.Result result = confirmationDialogController.showDialog(
+				messages.getString("CONFIRM_CUSTOM_FIELD_DELETION"),
+				MessageFormat.format(messages.getString("ARE_YOU_SURE_YOU_WANT_TO_DELETE_THE_FOLLOWING_CUSTOM_FIELDS"), tasksToDelete)
+		);
+		if (result == ConfirmationDialogController.Result.CONFIRMED)
+			for (CustomFieldAdapter customField : customfieldsToDeleteList)
+				dataManager.deleteCustomField(customField);
 	}
 }
